fcr_service/app/requirements.txt
pillow
torch
uvicorn
numpy
fastapi
facenet-pytorch
-f https://download.pytorch.org/whl/torch_stable.html